+++ src/main/java/com/codeborne/selenide/impl/SelenideElementProxy.java
@@ -28,7 +28,6 @@ import java.util.HashSet;
 import java.util.Set;
 
 import static com.codeborne.selenide.AssertionMode.SOFT;
-import static com.codeborne.selenide.impl.CommandGuard.executeWithTimeout;
 import static com.codeborne.selenide.logevents.ErrorsCollector.validateAssertionMode;
 import static com.codeborne.selenide.logevents.LogEvent.EventStatus.PASS;
 import static java.util.Arrays.asList;
@@ -136,8 +135,7 @@ class SelenideElementProxy<T extends SelenideElement> implements InvocationHandl
     do {
       try {
         if (isSelenideElementMethod(method)) {
-          return executeWithTimeout(timeoutMs + 5000, // just in case, give +5 seconds to avoid concurrent corner cases
-            () -> Commands.getInstance().execute(proxy, webElementSource, method.getName(), args));
+          return Commands.getInstance().execute(proxy, webElementSource, method.getName(), args);
         }
 
         return method.invoke(webElementSource.getWebElement(), args);